+++ internal/apps/dop/providers/issue/core/query/issueexcel/import.go
@@ -50,7 +50,7 @@ func ImportFile(r io.Reader, data *vars.DataForFulfill) error {
 	}
 
 	for _, h := range handlers {
-		if err := h.DecodeSheet(data, df); err != nil {
+		if err := h.ImportSheet(data, df); err != nil {
 			return fmt.Errorf("failed to decode sheet %q, err: %v", h.SheetName(), err)
 		}
 	}
