+++ internal/apps/dop/providers/issue/core/query/issueexcel/sheets/sheet_issue/import_vars_test.go
@@ -1,0 +1,34 @@
+// Copyright (c) 2021 Terminus, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package sheet_issue
+
+import (
+	"strings"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestNewIssueSheetColumnUUID(t *testing.T) {
+	uuid := NewIssueSheetColumnUUID("Common", "ID")
+	assert.Equal(t, uuid.String(), strings.Join([]string{"Common", "ID", "ID"}, issueSheetColumnUUIDSplitter))
+	uuid = NewIssueSheetColumnUUID("ID")
+	assert.Equal(t, uuid.String(), strings.Join([]string{"ID", "ID", "ID"}, issueSheetColumnUUIDSplitter))
+	uuid = NewIssueSheetColumnUUID("")
+	assert.Panicsf(t, func() { uuid.String() }, "uuid should not be empty")
+	uuid.AddPart("Common")
+	uuid.AddPart("ID")
+	assert.Equal(t, uuid.String(), strings.Join([]string{"Common", "ID", "ID"}, issueSheetColumnUUIDSplitter))
+}
